Big update adds the ability to send you emails
This big update, long desired by users, adds the ability to send you an email whenever malbait is tripped.
Qualifier: It might not work as I have not been able to test it properly. It's OK tho as this functionality is completely optional; you do not have to actually use it in this way.